requirements.txt
amqp==5.1.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.2.0
autopep8==2.0.2
Babel==2.14.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
bleach==6.1.0
bs4==0.0.1
cachelib==0.10.2
celery==5.2.7
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
checksumdir==1.2.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
contourpy==1.0.7
cryptography==39.0.2
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.6.0
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.3
fonttools==4.38.0
fqdn==1.5.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jupyter_core==5.6.1
jupyterlab_pygments==0.3.0
jwt==1.3.1
kiwisolver==1.4.4
kombu==5.2.4
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mistune==3.0.2
nest-asyncio==1.5.8
numpy==1.24.2
overrides==7.4.0
packaging==23.0
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
Pillow==9.4.0
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.38
psutil==5.9.7
psycopg2==2.9.5
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.7.1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
redis==4.5.4
referencing==0.32.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.16.2
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.4
stack-data==0.6.3
terminado==0.18.0
timedelta==2020.12.3
tinycss2==1.2.1
tomli==2.0.1
tornado==6.4
traitlets==5.14.1
types-python-dateutil==2.8.19.14
typing_extensions==4.5.0
tzdata==2023.4
uri-template==1.3.0
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==2.2.3
